done with middleware

+++ routes/web.php
@@ -17,24 +17,24 @@ use App\Http\Controllers\ListingController;
 // to get all listings from database
 Route::get('/', [ListingController::class,'index']);
 // to display create form
-Route::get('/listings/create', [ ListingController::class,'create']);
+Route::get('/listings/create', [ ListingController::class,'create'])->middleware('auth');
 // to store a single listing
-Route::post('/listings', [ListingController::class,'store']);
+Route::post('/listings', [ListingController::class,'store'])->middleware('auth');
 // to display a single listing
-Route::get('/listings/{id}', [ ListingController::class,'show']);
+Route::get('/listings/{id}', [ ListingController::class,'show'])->middleware('auth');
 // to display edit form
-Route::get('/listings/{id}/edit', [ ListingController::class,'edit']);
+Route::get('/listings/{id}/edit', [ ListingController::class,'edit'])->middleware('auth');
 // to update a single listing
-Route::put('/listings/{id}', [ ListingController::class,'update']);
+Route::put('/listings/{id}', [ ListingController::class,'update'])->middleware('auth');
 // to delete a single listing
-Route::delete('/listings/{id}', [ ListingController::class,'destroy']);
+Route::delete('/listings/{id}', [ ListingController::class,'destroy'])->middleware('auth');
 // to display register form
-Route::get('/register', [UserController::class,'create']);
+Route::get('/register', [UserController::class,'create'])->middleware('guest');
 // to store a single user
-Route::post('/register', [UserController::class,'store']);
+Route::post('/register', [UserController::class,'store'])->middleware('guest');
 // to logout
-Route::post('/logout', [UserController::class,'logout']);
+Route::post('/logout', [UserController::class,'logout'])->middleware('auth');
 // to display login form
-Route::get('/login', [UserController::class,'login']);
+Route::get('/login', [UserController::class,'login'])->name('login')->middleware('guest');;
 // to check login info
-Route::post('/login', [UserController::class,'loginHandler']);
+Route::post('/login', [UserController::class,'loginHandler'])->middleware('guest');
